Restoring cache.go file
